move things around; prepare for pt2
