+++ CMakeLists.txt
@@ -141,11 +141,14 @@ src/mrdsrg-spin-adapted/dsrg_mrpt_2nd.cc
 src/mrdsrg-spin-adapted/dsrg_mrpt_amp.cc
 src/mrdsrg-spin-adapted/dsrg_mrpt_comm.cc
 src/mrdsrg-spin-adapted/sadsrg.cc
+src/mrdsrg-spin-adapted/sadsrg_amps_analysis.cc
+src/mrdsrg-spin-adapted/sadsrg_block_labels.cc
 src/mrdsrg-spin-adapted/sadsrg_comm.cc
 src/mrdsrg-spin-adapted/sa_ldsrg2.cc
 src/mrdsrg-spin-adapted/sa_mrdsrg.cc
 src/mrdsrg-spin-adapted/sa_mrdsrg_amps.cc
 src/mrdsrg-spin-adapted/sa_mrdsrg_diis.cc
+src/mrdsrg-spin-adapted/sa_mrpt2.cc
 src/mrdsrg-spin-integrated/active_dsrgpt2.cc
 src/mrdsrg-spin-integrated/dsrg_mrpt2.cc
 src/mrdsrg-spin-integrated/dsrg_mrpt2_ms.cc
